DBConnection.ConnectionPool (db_connection v2.10.1)

The default connection pool.

The queueing algorithm is based on CoDel.

You're not supposed to call any functions on this pool directly, but only pass this as the value of the :pool option in functions such as DBConnection.start_link/2.

Functions

child_spec(init_arg)

Returns a specification to start this module under a supervisor.

See Supervisor.